fuel consuming clear

+++ UI/FuelBar.cs
@@ -9,9 +9,9 @@ public class FuelBar : MonoBehaviour
         _fuel = transform.GetChild(0).transform.GetChild(1).GetComponent<RectTransform>();
     }
 
-    public void SetFuel(short fuel, short maxFuel)
+    public void SetFuel(float fuel, float maxFuel)
     {
         if (fuel > maxFuel) return;
-        _fuel.localScale = new Vector3((float)fuel / maxFuel, 1, 1);
+        _fuel.localScale = new Vector3(fuel / maxFuel, 1, 1);
     }
 }
